Update .3
Sorry for the lack of updates, but we have gotten past the exam phase.

This update is the update that includes all of the restaurants in Starkville, MS!

Still going through and applying tags the restaurants to make specific types of food.

But hopefully now that I have gotten past some classes and the winter break would allow me to punch out a lot of this.